requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
openpyxl>=3.1.2
plotly>=5.18.0
litellm>=1.30.0
python-dotenv>=1.0.0
